Create test.py
test.py contains a function to get the variables' labels and a functions to normalize data within the PEEP level, PEEP trial or patient data (to be implemented).
Also it contains an experiment of the functions for 1 patient's data, as well as the conversion of the normalized data to a differentiation version of the data to understand the shift of the data.